add auth

+++ routes/api.php
@@ -22,7 +22,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::prefix('users')->group(function () {
     Route::post('/', [UserController::class, 'create']);
-    Route::put('/{user}', [UserController::class, 'update']);
-    Route::delete('/{user}', [UserController::class, 'delete']);
     Route::get('/', [UserController::class, 'list']);
+    Route::post('/auth/{user}', [UserController::class, 'authenticate']);
+
+    Route::middleware('auth:sanctum')->group(function () {
+        Route::middleware('auth:sanctum')->put('/{user}', [UserController::class, 'update']);
+        Route::middleware('auth:sanctum')->delete('/{user}', [UserController::class, 'delete']);
+    });
 });
